requirements.txt
fastapi==0.95.0
python-jose==3.3.0
pydantic==1.10.7
sklearn==0.0
numpy==1.21.5